update Exception (DatabaseException)

+++ src/main/java/com/devstarrk/dscommerce/services/ProductService.java
@@ -3,15 +3,16 @@ package com.devstarrk.dscommerce.services;
 import com.devstarrk.dscommerce.dto.ProductDTO;
 import com.devstarrk.dscommerce.entities.Product;
 import com.devstarrk.dscommerce.repositories.ProductRepository;
+import com.devstarrk.dscommerce.services.exceptions.DatabaseException;
 import com.devstarrk.dscommerce.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
-
-import java.util.List;
-import java.util.Optional;
 
 @Service
 public class ProductService {
@@ -38,14 +39,27 @@ public class ProductService {
     }
     @Transactional
     public ProductDTO update(Long id, ProductDTO dto){
-        Product entity = repository.getReferenceById(id);
-        copyDtoToEntity(dto, entity);
-        entity = repository.save(entity);
-        return new ProductDTO(entity);
+        try {
+            Product entity = repository.getReferenceById(id);
+            copyDtoToEntity(dto, entity);
+            entity = repository.save(entity);
+            return new ProductDTO(entity);
+        }
+        catch (EntityNotFoundException e){
+            throw new ResourceNotFoundException("Recurso nao encontrado");
+        }
     }
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id){
-        repository.deleteById(id);
+        if(!repository.existsById(id)){
+            throw new ResourceNotFoundException("Recurso nao encontrado");
+        }
+        try {
+            repository.deleteById(id);
+        }
+        catch (DataIntegrityViolationException e){
+            throw new DatabaseException("Falha na integridade referencial   ");
+        }
     }
 
     private void copyDtoToEntity(ProductDTO dto, Product entity) {
